refactor: assure main dependencies are initialized outside TSX components
