+++ src/data/context/RootContext/IRootContextProvider.ts
@@ -1,0 +1,7 @@
+import { PropsWithChildren } from 'react'
+import { IApiMiddleware } from '@src/middlewares'
+
+
+export interface IRootContextProvider extends PropsWithChildren {
+	apiMiddleware: IApiMiddleware
+}
